Write common function to read and create collection
ex : api/data/blog/post.json
insert in collection a basic structure

[
  {
    "collection": "posts",
    "timeStamp,
    "posts": []
  }
]

And read hooks for blog/posts.json "useReadGetAllPost"

unique code not deployable on another collection.
Refactored to read any kind of collection

to useReadCollection

The creation of a collection is still partial.
Feature to dev for the next commit. The complete create part.