Admin interface
Add post fully operational

--add post in BDD (json collection) and associate file image
upload in public directory.

--Next step read, update and delete data post

The data will not really be deleted but will be transferred to the same collection
inside a Trash folder.
The images will do the same.
Then you will have to write a script that cleans the data if it has three months of existence